Update romoff.asm
The introduced change allows you to disable, from the Pascal code level, the copying of font data from ROM to RAM when ROM is disabled. This process (unfortunate) overwrites the data in the area $E000..$E3FF, when we put in the resources that are loaded into this area. Using the definition of '{$DEFINE NOROMFONT}` you can disable the metastasis of font data from ROM to RAM, thus preserving, loaded resources.