test config example (Linux): update path to the binary.
The binary is expecting the libraries/units in subdirectories
of the directory from which it is run.
For testing, it is required then to make available what is needed,
- the binary of the compiler
- the libraries

To make it temporary (not sure what kind of changes will the compiler
undergo...), the configuration is prepared (and cleaned-up) in the
test/ directory.

Note:
For now, testing requires a hack: copying the binary and linking
required directories (another option could be to build it in
the main project directory, but that seems a bit messy... there is
already bin/ created for that).

It would be good to make it possible to specify the either
all needed paths, or the path to the directory under which
all of the directories with libs/units can be found. It could be
done either through a command-line option or an env. variable
(like MPLIBPATH).

Also, in unix-like platforms, the place where the binary executable
is stored should be somehow independent/configurable (either during
compilation or (rather also) with some configuration options or
env. variables. It would allow flexible installation and testing.