Use scoped enums
Types and scoping were introduced for TPass, TParameterPassingMethod, TCode65, TComputer, TCPU, TInterruptCode, TModifierBits, TIOCode, and TIOBits. This moved the identifiers from the global namespace to a local namespace. As a result, no prefixes/suffies like "__" or "i..." are required anymore.
The separation of the separate types for the bit-sets in method modifiers and I/O bits solve the problem of non-ordered and non-unique (e.g. two entries with ORD(x)=9) enum definitions.